// src/system_settings/dto/create-system_setting.dto.ts
export class CreateSystemSettingDto {
    abb: string;
    address: string;
    currency: string;
    email: string;
    fax: string;
    footer: string;
    image: string;
    map: string;
    mobile: string;
    name: string;
    sms_charges: string;
    stamp: string;
    timezone: string;
    title: string;
    website: string;
}